requirements.txt
flask== 1.0.2
pyaml==18.11.0
python-dateutil==2.8.0
SQLAlchemy==1.3.1
flask_sqlalchemy==2.3.2
flask-cors==3.0.2
pytest==4.0.1
config==0.4.2
boto==2.46.1
boto3==1.9.150
numpy==1.16.2
pandas==0.24.2
scikit-learn==0.20.3
sklearn==0.0
scipy==1.2.1
sympy==1.0
sqlalchemy==1.3.1
pickleshare==0.7.4
xgboost==0.90
